fix(app): NSNumber takes Float, no Float16 init
MLMultiArray with dataType .float16 still accepts NSNumber wrapping a
Float — the array converts on store. NSNumber has no Float16 init in
Swift 5.x, so use Float(1.0) / Float(0.0) for the attention_mask values.

+++ macOS/SideQuestApp/Models/EmbeddingGemmaModel.swift
@@ -87,10 +87,10 @@ actor EmbeddingGemmaModel {
       for i in 0..<seqLen {
         if i < realCount {
           idsArray[i] = NSNumber(value: Int32(tokenIds[i]))
-          maskArray[i] = NSNumber(value: Float16(1.0))
+          maskArray[i] = NSNumber(value: Float(1.0))
         } else {
           idsArray[i] = NSNumber(value: Int32(0))
-          maskArray[i] = NSNumber(value: Float16(0.0))
+          maskArray[i] = NSNumber(value: Float(0.0))
         }
       }
